# backend/api/docs/swagger.yaml
openapi: "3.0.0"
info:
  title: "Cubert Filesystem API"
  description: "Local filesystem exploration and management API"
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Development server
paths:
...
  /api/v1/filesystem/validate:
    post:
      tags:
        - "Filesystem"
      summary: "Validate path"
      description: "Validates if a path exists and is accessible"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path:
                  type: string
                  example: "/home/user"
      responses:
        "200":
          description: "Valid path"
          content:
            application/json:
              schema:
                type: object
                properties:
                  path:
                    type: string
                  valid:
                    type: boolean
        "400":
          description: "Invalid path"
...